api/requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
arrow==1.2.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.6
django-cors-headers==3.8.0
django-crontab==0.7.1
django-environ==0.5.0
django-filter==2.4.0
django-rest-knox==4.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-jwt
djangorestframework-simplejwt==4.4.0
djoser==2.1.0
docopt==0.6.2
filelock==3.0.12
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
mailjet-rest==1.3.4
Markdown==3.3.4
MarkupSafe==2.0.1
mysqlclient==2.0.3
mysql-connector-python==8.0.27
natsort==8.0.2
oauthlib==3.1.1
pipreqs==0.4.10
platformdirs==2.2.0
pony==0.7.14
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.4
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core
stripe==2.60.0
sqlparse==0.4.1
store==2020.2.10
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.7.2
xhtml2pdf==0.2.5
yarg==0.1.9
pillow  
drf-api-logger
django-nose 
coverage    
pytest-cov